requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
async_openai==0.0.21
discord.py==2.2.3
fakeyou==1.2.5
Flask==2.3.2
pydub==0.25.1
python_dateutil==2.8.2
SQLAlchemy==2.0.15
aiosqlite